+++ submitr/validators/__init__.py
@@ -3,3 +3,4 @@
 import submitr.validators.submitted_id_validator  # noqa
 import submitr.validators.duplicate_row_validator  # noqa
 import submitr.validators.file_set_count_validator  # noqa
+import submitr.validators.paired_read_validator  # noqa